# processor/lsmintervalprocessor: optimise marshalling