--- a/processor/lsmintervalprocessor/internal/merger/value_test.go
+++ b/processor/lsmintervalprocessor/internal/merger/value_test.go
@@ -62,11 +62,32 @@ func BenchmarkMerge(b *testing.B) {
 			dpLimit:     config.LimitConfig{MaxCardinality: 1},
 		},
 	} {
-		benchmarkWithTestdata(b, tc.name, tc.resLimit, tc.scopeLimit, tc.metricLimit, tc.dpLimit)
+		benchmarkMergeWithTestdata(b, tc.name, tc.resLimit, tc.scopeLimit, tc.metricLimit, tc.dpLimit)
 	}
 }
 
-func benchmarkWithTestdata(
+func BenchmarkMarshal(b *testing.B) {
+	for _, tc := range testCases {
+		dir := filepath.Join("../../testdata", tc)
+		md, err := golden.ReadMetrics(filepath.Join(dir, "input.yaml"))
+		require.NoError(b, err)
+		md.MarkReadOnly()
+		b.Run(tc, func(b *testing.B) {
+			maxLimit := config.LimitConfig{MaxCardinality: math.MaxInt64}
+			v := NewValue(maxLimit, maxLimit, maxLimit, maxLimit)
+			require.NoError(b, updateValueWithPMetrics(md, v))
+
+			//var buf []byte
+			for i := 0; i < b.N; i++ {
+				if _, err := v.Marshal(); err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+	}
+}
+
+func benchmarkMergeWithTestdata(
 	b *testing.B,
 	name string,
 	resLimit, scopeLimit, metricLimit, dpLimit config.LimitConfig,
@@ -78,19 +99,15 @@ func benchmarkWithTestdata(
 		md, err := golden.ReadMetrics(filepath.Join(dir, "input.yaml"))
 		require.NoError(b, err)
 		md.MarkReadOnly()
-		b.ResetTimer()
 		b.Run(name+"/"+tc, func(b *testing.B) {
+			to := NewValue(resLimit, scopeLimit, metricLimit, dpLimit)
+			from := NewValue(resLimit, scopeLimit, metricLimit, dpLimit)
+			require.NoError(b, updateValueWithPMetrics(md, from))
+
 			for i := 0; i < b.N; i++ {
-				b.StopTimer()
-				to := NewValue(resLimit, scopeLimit, metricLimit, dpLimit)
-				from := NewValue(resLimit, scopeLimit, metricLimit, dpLimit)
-				// Update from to have the pmetric structure
-				mdCopy := pmetric.NewMetrics()
-				md.CopyTo(mdCopy)
-				require.NoError(b, updateValueWithPMetrics(mdCopy, from))
-				b.StartTimer()
-				err = to.Merge(from)
-				require.NoError(b, err)
+				if err := to.Merge(from); err != nil {
+					b.Fatal(err)
+				}
 			}
 		})
 	}