Update axiomhq/hyperloglog and use the new AppendBinary method for encoding trackers. Update lsmintervalprocessor code to similarly append to a buffer to avoid creating and copying intermediate slices. Pool buffers to keep allocations to a minimum. If we can get https://github.com/open-telemetry/opentelemetry-collector/issues/12395 in, then we can shave off some more allocations.

There's quite a bit of variance in some of these benchmarks, but the results seem pretty promising:

```
goos: linux
goarch: amd64
pkg: github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics     
                                               │ /tmp/processor_old.txt │        /tmp/processor_new.txt        │
                                               │         sec/op         │    sec/op     vs base                │
Aggregation/sum_cumulative-16                              14.82µ ±  4%   14.73µ ±  5%        ~ (p=0.796 n=10)
Aggregation/sum_delta-16                                   16.51µ ±  7%   14.28µ ±  7%  -13.53% (p=0.000 n=10)
Aggregation/histogram_cumulative-16                        18.30µ ±  6%   16.59µ ±  4%   -9.32% (p=0.000 n=10)
Aggregation/histogram_delta-16                             19.17µ ±  8%   15.81µ ±  2%  -17.51% (p=0.000 n=10)
Aggregation/exphistogram_cumulative-16                     19.18µ ±  4%   16.15µ ±  5%  -15.78% (p=0.000 n=10)
Aggregation/exphistogram_delta-16                          17.89µ ±  6%   16.15µ ±  7%   -9.71% (p=0.000 n=10)
Aggregation/summary_enabled-16                             17.54µ ±  6%   14.30µ ±  7%  -18.46% (p=0.000 n=10)
Aggregation/summary_passthrough-16                         1.507µ ±  8%   1.190µ ±  7%  -21.07% (p=0.000 n=10)
AggregationWithOTTL/sum_cumulative-16                      24.33µ ± 12%   16.96µ ±  9%  -30.28% (p=0.000 n=10)
AggregationWithOTTL/sum_delta-16                           21.05µ ±  9%   17.77µ ± 11%  -15.57% (p=0.000 n=10)
AggregationWithOTTL/histogram_cumulative-16                23.28µ ± 13%   18.15µ ±  6%  -22.03% (p=0.000 n=10)
AggregationWithOTTL/histogram_delta-16                     23.82µ ± 11%   18.02µ ±  6%  -24.37% (p=0.000 n=10)
AggregationWithOTTL/exphistogram_cumulative-16             24.81µ ±  7%   19.25µ ±  7%  -22.41% (p=0.000 n=10)
AggregationWithOTTL/exphistogram_delta-16                  23.49µ ±  4%   19.51µ ±  2%  -16.91% (p=0.000 n=10)
AggregationWithOTTL/summary_enabled-16                     21.49µ ±  4%   17.82µ ±  5%  -17.06% (p=0.000 n=10)
AggregationWithOTTL/summary_passthrough-16                 1.572µ ± 10%   1.290µ ±  4%  -17.97% (p=0.000 n=10)
geomean                                                    14.62µ         12.09µ        -17.30%

                                               │ /tmp/processor_old.txt │        /tmp/processor_new.txt        │
                                               │          B/op          │     B/op      vs base                │
Aggregation/sum_cumulative-16                             13.06Ki ±  4%   11.65Ki ± 1%  -10.76% (p=0.000 n=10)
Aggregation/sum_delta-16                                  12.74Ki ±  1%   11.53Ki ± 1%   -9.48% (p=0.000 n=10)
Aggregation/histogram_cumulative-16                       14.82Ki ± 28%   13.14Ki ± 2%  -11.33% (p=0.000 n=10)
Aggregation/histogram_delta-16                            18.63Ki ± 23%   13.10Ki ± 3%  -29.69% (p=0.000 n=10)
Aggregation/exphistogram_cumulative-16                    14.57Ki ±  2%   13.19Ki ± 2%   -9.49% (p=0.000 n=10)
Aggregation/exphistogram_delta-16                         14.47Ki ±  1%   13.27Ki ± 1%   -8.30% (p=0.000 n=10)
Aggregation/summary_enabled-16                            12.99Ki ±  2%   11.69Ki ± 1%  -10.04% (p=0.000 n=10)
Aggregation/summary_passthrough-16                        1.342Ki ±  1%   1.327Ki ± 1%   -1.13% (p=0.000 n=10)
AggregationWithOTTL/sum_cumulative-16                     15.47Ki ±  3%   13.95Ki ± 1%   -9.86% (p=0.000 n=10)
AggregationWithOTTL/sum_delta-16                          15.21Ki ±  3%   14.06Ki ± 2%   -7.53% (p=0.000 n=10)
AggregationWithOTTL/histogram_cumulative-16               16.90Ki ±  4%   15.49Ki ± 2%   -8.33% (p=0.000 n=10)
AggregationWithOTTL/histogram_delta-16                    16.83Ki ±  5%   15.47Ki ± 1%   -8.08% (p=0.000 n=10)
AggregationWithOTTL/exphistogram_cumulative-16            17.03Ki ±  3%   15.65Ki ± 2%   -8.10% (p=0.000 n=10)
AggregationWithOTTL/exphistogram_delta-16                 17.04Ki ±  0%   15.60Ki ± 2%   -8.42% (p=0.000 n=10)
AggregationWithOTTL/summary_enabled-16                    15.49Ki ±  1%   14.03Ki ± 1%   -9.42% (p=0.000 n=10)
AggregationWithOTTL/summary_passthrough-16                1.345Ki ±  1%   1.326Ki ± 1%   -1.42% (p=0.000 n=10)
geomean                                                   11.28Ki         10.19Ki        -9.68%

                                               │ /tmp/processor_old.txt │       /tmp/processor_new.txt       │
                                               │       allocs/op        │ allocs/op   vs base                │
Aggregation/sum_cumulative-16                               137.0 ±  0%   126.0 ± 0%   -8.03% (p=0.000 n=10)
Aggregation/sum_delta-16                                    138.0 ±  0%   127.0 ± 0%   -7.97% (p=0.000 n=10)
Aggregation/histogram_cumulative-16                         145.0 ± 37%   134.0 ± 0%   -7.59% (p=0.000 n=10)
Aggregation/histogram_delta-16                              198.5 ± 26%   135.0 ± 0%  -31.99% (p=0.000 n=10)
Aggregation/exphistogram_cumulative-16                      149.0 ±  0%   138.0 ± 0%   -7.38% (p=0.000 n=10)
Aggregation/exphistogram_delta-16                           150.0 ±  0%   139.0 ± 0%   -7.33% (p=0.000 n=10)
Aggregation/summary_enabled-16                              140.0 ±  0%   129.0 ± 0%   -7.86% (p=0.000 n=10)
Aggregation/summary_passthrough-16                          34.00 ±  6%   33.00 ± 0%   -2.94% (p=0.000 n=10)
AggregationWithOTTL/sum_cumulative-16                       164.0 ±  0%   153.0 ± 0%   -6.71% (p=0.000 n=10)
AggregationWithOTTL/sum_delta-16                            165.0 ±  0%   154.0 ± 0%   -6.67% (p=0.000 n=10)
AggregationWithOTTL/histogram_cumulative-16                 172.0 ±  0%   161.0 ± 0%   -6.40% (p=0.000 n=10)
AggregationWithOTTL/histogram_delta-16                      173.0 ±  0%   162.0 ± 0%   -6.36% (p=0.000 n=10)
AggregationWithOTTL/exphistogram_cumulative-16              176.0 ±  0%   165.0 ± 0%   -6.25% (p=0.000 n=10)
AggregationWithOTTL/exphistogram_delta-16                   177.0 ±  0%   166.0 ± 0%   -6.21% (p=0.000 n=10)
AggregationWithOTTL/summary_enabled-16                      167.0 ±  0%   156.0 ± 0%   -6.59% (p=0.000 n=10)
AggregationWithOTTL/summary_passthrough-16                  34.00 ±  6%   33.00 ± 0%   -2.94% (p=0.000 n=10)
geomean                                                     131.8         120.8        -8.34%
```