@axw 2025-02-18
In https://github.com/elastic/opentelemetry-collector-components/pull/290/commits/7ce291a4b6f4e9b1ce0b7c54d95db35cab527b9d I've made some more targeted optimisations in the merger and limits packages:
 - Trackers no longer encode a version, that's encoded at the top level, i.e. in merger.Value. This is a breaking change at the moment, we'll need to figure out how if we need to maintain backwards compatibility at this early stage.
 - Use a varint for encoding the observed count. This will use less space at low throughput.
 - Use just 2 bytes for encoding the length of the HLL++ sketch, which is sufficient for the precision we're using
 - Use 4 bytes for encoding the length of all trackers (4GB should be more than enough!)
 - Stop encoding the length of each individual tracker, instead just unmarshal each one in series and keep track of how many bytes were consumed for each one
 - Stop including the estimated trackers length and encoded protobuf length in the slices.Grow call, which apparently makes things slower

```
goos: linux
goarch: amd64
pkg: github.com/elastic/opentelemetry-collector-components/processor/lsmintervalprocessor/internal/merger
cpu: AMD Ryzen 7 PRO 5850U with Radeon Graphics     
                                                            │ /tmp/old.txt  │             /tmp/new.txt             │
                                                            │    sec/op     │    sec/op     vs base                │
Merge/without_overflow/sum_cumulative_overflow-16              1.044µ ± 12%   1.028µ ±  7%        ~ (p=0.353 n=10)
Merge/without_overflow/sum_delta_overflow-16                   1.150µ ±  9%   1.034µ ±  7%  -10.05% (p=0.002 n=10)
Merge/without_overflow/histogram_cumulative_overflow-16        1.322µ ± 16%   1.040µ ± 10%  -21.29% (p=0.000 n=10)
Merge/without_overflow/histogram_delta_overflow-16             1.363µ ±  8%   1.151µ ±  3%  -15.56% (p=0.000 n=10)
Merge/without_overflow/exphistogram_cumulative_overflow-16     1.014µ ± 14%   1.040µ ±  9%        ~ (p=0.644 n=10)
Merge/without_overflow/exphistogram_delta_overflow-16          1.215µ ± 10%   1.147µ ±  4%   -5.56% (p=0.041 n=10)
Merge/with_overflow/sum_cumulative_overflow-16                 709.4n ± 11%   578.8n ±  9%  -18.41% (p=0.000 n=10)
Merge/with_overflow/sum_delta_overflow-16                      646.4n ± 14%   567.5n ± 11%  -12.21% (p=0.002 n=10)
Merge/with_overflow/histogram_cumulative_overflow-16           636.5n ±  9%   589.6n ± 10%        ~ (p=0.089 n=10)
Merge/with_overflow/histogram_delta_overflow-16                736.9n ±  6%   616.1n ±  9%  -16.39% (p=0.000 n=10)
Merge/with_overflow/exphistogram_cumulative_overflow-16        648.3n ±  5%   538.8n ±  9%  -16.88% (p=0.000 n=10)
Merge/with_overflow/exphistogram_delta_overflow-16             922.2n ± 17%   661.5n ± 10%  -28.26% (p=0.000 n=10)
Marshal/sum_cumulative_overflow-16                            1213.0n ± 15%   819.2n ±  9%  -32.46% (p=0.000 n=10)
Marshal/sum_delta_overflow-16                                 1160.5n ±  7%   847.0n ±  6%  -27.02% (p=0.000 n=10)
Marshal/histogram_cumulative_overflow-16                       1.219µ ±  9%   1.170µ ±  9%   -4.02% (p=0.000 n=10)
Marshal/histogram_delta_overflow-16                            1.318µ ± 45%   1.054µ ± 10%  -20.03% (p=0.000 n=10)
Marshal/exphistogram_cumulative_overflow-16                    1.748µ ±  5%   1.618µ ±  3%   -7.44% (p=0.000 n=10)
Marshal/exphistogram_delta_overflow-16                         1.990µ ± 28%   1.613µ ±  2%  -18.92% (p=0.000 n=10)
AppendBinaryReuseBuffer/sum_cumulative_overflow-16                            718.3n ±  8%
AppendBinaryReuseBuffer/sum_delta_overflow-16                                 710.3n ±  8%
AppendBinaryReuseBuffer/histogram_cumulative_overflow-16                      816.3n ±  9%
AppendBinaryReuseBuffer/histogram_delta_overflow-16                           857.3n ±  6%
AppendBinaryReuseBuffer/exphistogram_cumulative_overflow-16                   1.308µ ±  7%
AppendBinaryReuseBuffer/exphistogram_delta_overflow-16                        1.436µ ± 10%
geomean                                                        1.057µ         907.2n        -15.01%

                                                            │  /tmp/old.txt  │              /tmp/new.txt              │
                                                            │      B/op      │     B/op      vs base                  │
Merge/without_overflow/sum_cumulative_overflow-16               0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/sum_delta_overflow-16                    24.00 ± 0%       24.00 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/histogram_cumulative_overflow-16         0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/histogram_delta_overflow-16              24.00 ± 0%       24.00 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/exphistogram_cumulative_overflow-16      0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/exphistogram_delta_overflow-16           24.00 ± 0%       24.00 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/sum_cumulative_overflow-16                  0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/sum_delta_overflow-16                       8.000 ± 0%       8.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/histogram_cumulative_overflow-16            0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/histogram_delta_overflow-16                 8.000 ± 0%       8.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/exphistogram_cumulative_overflow-16         0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/exphistogram_delta_overflow-16              8.000 ± 0%       8.000 ± 0%        ~ (p=1.000 n=10) ¹
Marshal/sum_cumulative_overflow-16                             1088.0 ± 0%       712.0 ± 0%  -34.56% (p=0.000 n=10)
Marshal/sum_delta_overflow-16                                  1088.0 ± 0%       712.0 ± 0%  -34.56% (p=0.000 n=10)
Marshal/histogram_cumulative_overflow-16                      1.719Ki ± 0%     1.352Ki ± 0%  -21.36% (p=0.000 n=10)
Marshal/histogram_delta_overflow-16                           1.719Ki ± 0%     1.289Ki ± 0%  -25.00% (p=0.000 n=10)
Marshal/exphistogram_cumulative_overflow-16                   1.688Ki ± 0%     1.289Ki ± 0%  -23.61% (p=0.000 n=10)
Marshal/exphistogram_delta_overflow-16                        1.703Ki ± 0%     1.305Ki ± 0%  -23.39% (p=0.000 n=10)
AppendBinaryReuseBuffer/sum_cumulative_overflow-16                               320.0 ± 0%
AppendBinaryReuseBuffer/sum_delta_overflow-16                                    320.0 ± 0%
AppendBinaryReuseBuffer/histogram_cumulative_overflow-16                         640.0 ± 0%
AppendBinaryReuseBuffer/histogram_delta_overflow-16                              640.0 ± 0%
AppendBinaryReuseBuffer/exphistogram_cumulative_overflow-16                      800.0 ± 0%
AppendBinaryReuseBuffer/exphistogram_delta_overflow-16                           816.0 ± 0%
geomean                                                                    ²                 -10.08%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                            │ /tmp/old.txt │             /tmp/new.txt             │
                                                            │  allocs/op   │ allocs/op   vs base                  │
Merge/without_overflow/sum_cumulative_overflow-16             0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/sum_delta_overflow-16                  3.000 ± 0%     3.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/histogram_cumulative_overflow-16       0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/histogram_delta_overflow-16            3.000 ± 0%     3.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/exphistogram_cumulative_overflow-16    0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/without_overflow/exphistogram_delta_overflow-16         3.000 ± 0%     3.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/sum_cumulative_overflow-16                0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/sum_delta_overflow-16                     1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/histogram_cumulative_overflow-16          0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/histogram_delta_overflow-16               1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/exphistogram_cumulative_overflow-16       0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=10) ¹
Merge/with_overflow/exphistogram_delta_overflow-16            1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=10) ¹
Marshal/sum_cumulative_overflow-16                            9.000 ± 0%     4.000 ± 0%  -55.56% (p=0.000 n=10)
Marshal/sum_delta_overflow-16                                 9.000 ± 0%     4.000 ± 0%  -55.56% (p=0.000 n=10)
Marshal/histogram_cumulative_overflow-16                      9.000 ± 0%     4.000 ± 0%  -55.56% (p=0.000 n=10)
Marshal/histogram_delta_overflow-16                           9.000 ± 0%     4.000 ± 0%  -55.56% (p=0.000 n=10)
Marshal/exphistogram_cumulative_overflow-16                   15.00 ± 0%     10.00 ± 0%  -33.33% (p=0.000 n=10)
Marshal/exphistogram_delta_overflow-16                        15.00 ± 0%     10.00 ± 0%  -33.33% (p=0.000 n=10)
AppendBinaryReuseBuffer/sum_cumulative_overflow-16                           1.000 ± 0%
AppendBinaryReuseBuffer/sum_delta_overflow-16                                1.000 ± 0%
AppendBinaryReuseBuffer/histogram_cumulative_overflow-16                     1.000 ± 0%
AppendBinaryReuseBuffer/histogram_delta_overflow-16                          1.000 ± 0%
AppendBinaryReuseBuffer/exphistogram_cumulative_overflow-16                  7.000 ± 0%
AppendBinaryReuseBuffer/exphistogram_delta_overflow-16                       7.000 ± 0%
geomean                                                                  ²               -20.17%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```